school_management_system/requirements.txt
amqp==5.3.1
openpyxl==3.1.5
django-cors-headers==4.9.0
asgiref==3.9.1
billiard==4.2.2
Brotli==1.1.0
celery==5.5.3
cffi==2.0.0
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cssselect2==0.8.0
Django==5.2.6
dj-database-url==2.2.0
django-crispy-forms==2.3
django-filter==25.1
django-guardian==3.2.0
django-phonenumber-field==8.1.0
django-rest-framework==0.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
fonttools==4.60.0
kombu==5.5.4
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pycparser==2.23
pydyf==0.11.0
PyJWT==2.10.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
redis==6.4.0
requests==2.31.0
six==1.17.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
weasyprint==66.0
webencodings==0.5.1
zopfli==0.2.3.post1
